+++ tests/testthat/test-QueryNamespace.R
@@ -69,25 +69,31 @@ test_that("test executeSql function", {
 test_that("create helper function works", {
   expect_error(createQueryNamespace())
 
-  qns <- createQueryNamespace(connectionDetails = connectionDetails,
-                              usePooledConnection = FALSE,
-                              tableSpecification = loadResultsDataModelSpecifications("settings/resultsDataModelSpecification.csv"),
-                              resultModelSpecificationPath = NULL,
-                              tablePrefix = "",
-                              snakeCaseToCamelCase = TRUE,
-                              databaseSchema = "main")
+  qns <- createQueryNamespace(
+    connectionDetails = connectionDetails,
+    usePooledConnection = FALSE,
+    tableSpecification = loadResultsDataModelSpecifications("settings/resultsDataModelSpecification.csv"),
+    resultModelSpecificationPath = NULL,
+    tablePrefix = "",
+    snakeCaseToCamelCase = TRUE,
+    databaseSchema = "main"
+  )
 
   vars <- qns$getVars()
   expect_true("databaseSchema" %in% names(vars))
 
-  qns <- createQueryNamespace(connectionDetails = connectionDetails,
-                              usePooledConnection = TRUE,
-                              tableSpecification = NULL,
-                              resultModelSpecificationPath = c("settings/resultsDataModelSpecification.csv",
-                                                               "settings/testSchemaDef.csv"),
-                              tablePrefix = "",
-                              snakeCaseToCamelCase = TRUE,
-                              databaseSchema = "main")
+  qns <- createQueryNamespace(
+    connectionDetails = connectionDetails,
+    usePooledConnection = TRUE,
+    tableSpecification = NULL,
+    resultModelSpecificationPath = c(
+      "settings/resultsDataModelSpecification.csv",
+      "settings/testSchemaDef.csv"
+    ),
+    tablePrefix = "",
+    snakeCaseToCamelCase = TRUE,
+    databaseSchema = "main"
+  )
 
   vars <- qns$getVars()
 
@@ -101,22 +107,26 @@ test_that("create helper function works", {
   expect_true("test_table_1" %in% names(vars))
 
   expect_error(
-    createQueryNamespace(connectionDetails = connectionDetails,
-                         usePooledConnection = TRUE,
-                         tableSpecification = NULL,
-                         resultModelSpecificationPath = c("fileDoesNotExist"),
-                         tablePrefix = "",
-                         snakeCaseToCamelCase = TRUE,
-                         databaseSchema = "main")
+    createQueryNamespace(
+      connectionDetails = connectionDetails,
+      usePooledConnection = TRUE,
+      tableSpecification = NULL,
+      resultModelSpecificationPath = c("fileDoesNotExist"),
+      tablePrefix = "",
+      snakeCaseToCamelCase = TRUE,
+      databaseSchema = "main"
+    )
   )
 
   expect_error(
-     createQueryNamespace(connectionDetails = NULL,
-                          usePooledConnection = FALSE,
-                          tableSpecification = NULL,
-                          resultModelSpecificationPath = c("settings/resultsDataModelSpecification.csv"),
-                          tablePrefix = "",
-                          snakeCaseToCamelCase = TRUE,
-                          databaseSchema = "main")
+    createQueryNamespace(
+      connectionDetails = NULL,
+      usePooledConnection = FALSE,
+      tableSpecification = NULL,
+      resultModelSpecificationPath = c("settings/resultsDataModelSpecification.csv"),
+      tablePrefix = "",
+      snakeCaseToCamelCase = TRUE,
+      databaseSchema = "main"
+    )
   )
 })